reducers: extract fragmentReducer to its own file
